fiddling with performance issues and optimization.
I moved a copy of the bunch test into the tests directory,
changed it from a bench to a test, and marked it ignored.
I wanted to be able to run it the way I used to, by itself,
with 100 thousand as the basic size.

I found the stuff in the Cargo.toml manifest docs to configure
optimization profile.  I configured profile.test to build with
opt-level 3.

I ran the test, and it was really slow, as if the optimizer
wasn't doing anything at all.

So I realized the main difference here is that the test is
no longer in the same crate as the library itself.

I moved another copy of the test back into the same crate as
the library and compiled it with optimization.  Now it's back
to being around 15x as fast.

Proceeding with the conjecture that this is all about cross-crate
inlining, I went digging and found this:

https://internals.rust-lang.org/t/when-should-i-use-inline/598/2

So I changed the profile.test to compile with lto, following the
idea that this would basically allow anything to be inlined at
link time whether I put inline attributes on my functions or
not.

Rebuild, run the test, and now it crashes with a seg fault.

Still digging...