redo all the error handling.
http://blog.burntsushi.net/rust-error-handling/

Instead of using io::Result everywhere, and using io::ErrorKind::Other
with strings for all the cases that don't fit, create a custom error
Result type and a custom error type LsmError.  Io is one of the
options in the custom Result type, but there are specific cases
for other kinds of errors as well.  use impl From to allow automatic
wrapping of io:Error and PoisonError as an LsmError so that try!
is clean everywhere.